docs(encrypted-archive): move example codes for Node.js and React Native into same directory `examples`
+ `examples/*` -> `examples/node/*`
+ `example/*`  -> `examples/react-native/*`
