+++ packages/encrypted-archive/examples/node/cleartext.txt
