+++ packages/encrypted-archive/examples/node/index.js
