+++ packages/encrypted-archive/examples/node/package.json
@@ -1,7 +1,7 @@
 {
   "private": true,
   "dependencies": {
-    "@sounisi5011/encrypted-archive": "file:.."
+    "@sounisi5011/encrypted-archive": "file:../.."
   },
   "engines": {
     "node": "^12.17.x || 14.x || 15.x || 16.x || 17.x"
