+++ packages/encrypted-archive/examples/node/stream.js
