+++ packages/encrypted-archive/examples/react-native/android/app/build.gradle
