+++ packages/encrypted-archive/examples/react-native/android/app/debug.keystore
