+++ packages/encrypted-archive/examples/react-native/android/app/proguard-rules.pro
