+++ packages/encrypted-archive/examples/react-native/android/app/src/debug/AndroidManifest.xml
