+++ packages/encrypted-archive/examples/react-native/android/app/src/main/AndroidManifest.xml
