+++ packages/encrypted-archive/examples/react-native/android/app/src/main/java/com/example/sounisi5011encryptedarchive/MainActivity.java
