+++ packages/encrypted-archive/examples/react-native/android/app/src/main/res/mipmap-mdpi/ic_launcher.png
