+++ packages/encrypted-archive/examples/react-native/android/app/src/main/res/mipmap-xxhdpi/ic_launcher_round.png
