+++ packages/encrypted-archive/examples/react-native/android/app/src/main/res/mipmap-xxxhdpi/ic_launcher.png
