+++ packages/encrypted-archive/examples/react-native/android/app/src/main/res/values/styles.xml
