+++ packages/encrypted-archive/examples/react-native/android/gradle.properties
