+++ packages/encrypted-archive/examples/react-native/android/gradle/wrapper/gradle-wrapper.properties
