+++ packages/encrypted-archive/examples/react-native/android/gradlew
