+++ packages/encrypted-archive/examples/react-native/android/settings.gradle
