+++ packages/encrypted-archive/examples/react-native/app.json
