+++ packages/encrypted-archive/examples/react-native/babel.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const pak = require('../package.json');
+const pak = require('../../package.json');
 
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
