+++ packages/encrypted-archive/examples/react-native/ios/EncryptedArchiveExample.xcodeproj/xcshareddata/xcschemes/EncryptedArchiveExample.xcscheme
