+++ packages/encrypted-archive/examples/react-native/ios/EncryptedArchiveExample.xcworkspace/contents.xcworkspacedata
