+++ packages/encrypted-archive/examples/react-native/ios/EncryptedArchiveExample.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
