+++ packages/encrypted-archive/examples/react-native/ios/EncryptedArchiveExample/Images.xcassets/AppIcon.appiconset/Contents.json
