+++ packages/encrypted-archive/examples/react-native/ios/EncryptedArchiveExample/Images.xcassets/Contents.json
