+++ packages/encrypted-archive/examples/react-native/ios/EncryptedArchiveExample/Info.plist
