+++ packages/encrypted-archive/examples/react-native/ios/File.swift
