+++ packages/encrypted-archive/examples/react-native/metro.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const blacklist = require('metro-config/src/defaults/blacklist');
 const escape = require('escape-string-regexp');
-const pak = require('../package.json');
+const pak = require('../../package.json');
 
 const root = path.resolve(__dirname, '..');
