+++ packages/encrypted-archive/examples/react-native/package.json
