+++ packages/encrypted-archive/examples/react-native/src/App.tsx
